// src/api/generated/model/jiraUserResponseDto.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Dev Blackbox API
 * OpenAPI spec version: 0.1.0
 */

export interface JiraUserResponseDto {
  id: number;
  account_id: string;
  active: boolean;
  display_name: string;
  email_address: string;
  url: string;
  project: string | null;
  user_id: number | null;
  created_at: string;
  updated_at: string;
}